Move Store dependencies to store
